requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1

openai==1.54.3
chromadb==0.5.18
tiktoken==0.8.0
tenacity==9.0.0

python-gitlab==4.13.0
gitpython==3.1.43

httpx==0.27.2
python-multipart==0.0.12
slowapi==0.1.9

pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.4
